New branch with exception

+++ src/ua/ithillel/logger/FileLoggerConfiguration.java
@@ -2,5 +2,5 @@ package ua.ithillel.logger;
 
 import java.io.File;
 
-public record FileLoggerConfiguration(File FILE, LoggingLevel LOGGING_LEVEL, int MAX_SIZE, String FORMAT) {
+public record FileLoggerConfiguration(File FILE, LoggingLevel LOGGING_LEVEL, long MAX_SIZE, String FORMAT) {
 }
